build: add enable_remote_module build flag
