+++ BUILD.gn
@@ -579,6 +579,17 @@ source_set("electron_lib") {
     ]
   }
 
+  if (enable_remote_module) {
+    sources += [
+      "shell/common/api/remote/object_life_monitor.cc",
+      "shell/common/api/remote/object_life_monitor.h",
+      "shell/common/api/remote/remote_callback_freer.cc",
+      "shell/common/api/remote/remote_callback_freer.h",
+      "shell/common/api/remote/remote_object_freer.cc",
+      "shell/common/api/remote/remote_object_freer.h",
+    ]
+  }
+
   if (enable_desktop_capturer) {
     if (is_component_build && !is_linux) {
       # On windows the implementation relies on unexported
