+++ filenames.auto.gni
@@ -133,20 +133,19 @@ auto_filenames = {
     "lib/common/api/module-list.ts",
     "lib/common/api/native-image.js",
     "lib/common/api/shell.js",
-    "lib/common/buffer-utils.ts",
     "lib/common/clipboard-utils.ts",
     "lib/common/crash-reporter.js",
     "lib/common/define-properties.ts",
     "lib/common/electron-binding-setup.ts",
     "lib/common/error-utils.ts",
-    "lib/common/is-promise.ts",
+    "lib/common/remote/buffer-utils.ts",
+    "lib/common/remote/is-promise.ts",
     "lib/common/web-view-methods.ts",
     "lib/renderer/api/crash-reporter.js",
     "lib/renderer/api/desktop-capturer.ts",
     "lib/renderer/api/ipc-renderer.ts",
     "lib/renderer/api/remote.js",
     "lib/renderer/api/web-frame.ts",
-    "lib/renderer/callbacks-registry.ts",
     "lib/renderer/chrome-api.ts",
     "lib/renderer/content-scripts-injector.ts",
     "lib/renderer/extensions/event.ts",
@@ -156,6 +155,7 @@ auto_filenames = {
     "lib/renderer/inspector.ts",
     "lib/renderer/ipc-renderer-internal-utils.ts",
     "lib/renderer/ipc-renderer-internal.ts",
+    "lib/renderer/remote/callbacks-registry.ts",
     "lib/renderer/security-warnings.ts",
     "lib/renderer/web-frame-init.ts",
     "lib/renderer/web-view/guest-view-internal.ts",
@@ -258,7 +258,8 @@ auto_filenames = {
     "lib/browser/ipc-main-internal-utils.ts",
     "lib/browser/ipc-main-internal.ts",
     "lib/browser/navigation-controller.js",
-    "lib/browser/objects-registry.js",
+    "lib/browser/remote/objects-registry.js",
+    "lib/browser/remote/server.js",
     "lib/browser/rpc-server.js",
     "lib/browser/utils.ts",
     "lib/common/api/clipboard.js",
@@ -266,15 +267,15 @@ auto_filenames = {
     "lib/common/api/module-list.ts",
     "lib/common/api/native-image.js",
     "lib/common/api/shell.js",
-    "lib/common/buffer-utils.ts",
     "lib/common/clipboard-utils.ts",
     "lib/common/crash-reporter.js",
     "lib/common/define-properties.ts",
     "lib/common/electron-binding-setup.ts",
     "lib/common/error-utils.ts",
     "lib/common/init.ts",
-    "lib/common/is-promise.ts",
     "lib/common/parse-features-string.js",
+    "lib/common/remote/buffer-utils.ts",
+    "lib/common/remote/is-promise.ts",
     "lib/common/reset-search-paths.ts",
     "lib/common/web-view-methods.ts",
     "lib/renderer/ipc-renderer-internal-utils.ts",
@@ -291,14 +292,14 @@ auto_filenames = {
     "lib/common/api/module-list.ts",
     "lib/common/api/native-image.js",
     "lib/common/api/shell.js",
-    "lib/common/buffer-utils.ts",
     "lib/common/clipboard-utils.ts",
     "lib/common/crash-reporter.js",
     "lib/common/define-properties.ts",
     "lib/common/electron-binding-setup.ts",
     "lib/common/error-utils.ts",
     "lib/common/init.ts",
-    "lib/common/is-promise.ts",
+    "lib/common/remote/buffer-utils.ts",
+    "lib/common/remote/is-promise.ts",
     "lib/common/reset-search-paths.ts",
     "lib/common/web-view-methods.ts",
     "lib/renderer/api/crash-reporter.js",
@@ -308,7 +309,6 @@ auto_filenames = {
     "lib/renderer/api/module-list.ts",
     "lib/renderer/api/remote.js",
     "lib/renderer/api/web-frame.ts",
-    "lib/renderer/callbacks-registry.ts",
     "lib/renderer/chrome-api.ts",
     "lib/renderer/content-scripts-injector.ts",
     "lib/renderer/extensions/event.ts",
@@ -319,6 +319,7 @@ auto_filenames = {
     "lib/renderer/inspector.ts",
     "lib/renderer/ipc-renderer-internal-utils.ts",
     "lib/renderer/ipc-renderer-internal.ts",
+    "lib/renderer/remote/callbacks-registry.ts",
     "lib/renderer/security-warnings.ts",
     "lib/renderer/web-frame-init.ts",
     "lib/renderer/web-view/guest-view-internal.ts",
@@ -341,14 +342,14 @@ auto_filenames = {
     "lib/common/api/module-list.ts",
     "lib/common/api/native-image.js",
     "lib/common/api/shell.js",
-    "lib/common/buffer-utils.ts",
     "lib/common/clipboard-utils.ts",
     "lib/common/crash-reporter.js",
     "lib/common/define-properties.ts",
     "lib/common/electron-binding-setup.ts",
     "lib/common/error-utils.ts",
     "lib/common/init.ts",
-    "lib/common/is-promise.ts",
+    "lib/common/remote/buffer-utils.ts",
+    "lib/common/remote/is-promise.ts",
     "lib/common/reset-search-paths.ts",
     "lib/renderer/api/crash-reporter.js",
     "lib/renderer/api/desktop-capturer.ts",
@@ -357,9 +358,9 @@ auto_filenames = {
     "lib/renderer/api/module-list.ts",
     "lib/renderer/api/remote.js",
     "lib/renderer/api/web-frame.ts",
-    "lib/renderer/callbacks-registry.ts",
     "lib/renderer/ipc-renderer-internal-utils.ts",
     "lib/renderer/ipc-renderer-internal.ts",
+    "lib/renderer/remote/callbacks-registry.ts",
     "lib/renderer/webpack-provider.ts",
     "lib/worker/init.js",
     "package.json",
