+++ filenames.gni
@@ -440,12 +440,6 @@ filenames = {
     "shell/common/api/features.cc",
     "shell/common/api/locker.cc",
     "shell/common/api/locker.h",
-    "shell/common/api/object_life_monitor.cc",
-    "shell/common/api/object_life_monitor.h",
-    "shell/common/api/remote_callback_freer.cc",
-    "shell/common/api/remote_callback_freer.h",
-    "shell/common/api/remote_object_freer.cc",
-    "shell/common/api/remote_object_freer.h",
     "shell/common/asar/archive.cc",
     "shell/common/asar/archive.h",
     "shell/common/asar/asar_util.cc",
