+++ lib/browser/remote/objects-registry.js
