+++ lib/common/remote/buffer-utils.ts
