+++ lib/common/remote/is-promise.ts
