+++ lib/renderer/api/remote.js
@@ -2,10 +2,10 @@
 
 const v8Util = process.electronBinding('v8_util')
 
-const { CallbacksRegistry } = require('@electron/internal/renderer/callbacks-registry')
-const bufferUtils = require('@electron/internal/common/buffer-utils')
+const { CallbacksRegistry } = require('@electron/internal/renderer/remote/callbacks-registry')
+const bufferUtils = require('@electron/internal/common/remote/buffer-utils')
 const errorUtils = require('@electron/internal/common/error-utils')
-const { isPromise } = require('@electron/internal/common/is-promise')
+const { isPromise } = require('@electron/internal/common/remote/is-promise')
 const { ipcRendererInternal } = require('@electron/internal/renderer/ipc-renderer-internal')
 
 const callbacksRegistry = new CallbacksRegistry()
