+++ lib/renderer/remote/callbacks-registry.ts
