+++ spec-main/api-callbacks-registry-spec.ts
@@ -1,7 +1,10 @@
 import { expect } from 'chai'
-import { CallbacksRegistry } from '../lib/renderer/callbacks-registry'
+import { CallbacksRegistry } from '../lib/renderer/remote/callbacks-registry'
+import { ifdescribe } from './spec-helpers';
 
-describe('CallbacksRegistry module', () => {
+const features = process.electronBinding('features')
+
+ifdescribe(features.isRemoteModuleEnabled())('CallbacksRegistry module', () => {
   let registry: CallbacksRegistry
 
   beforeEach(() => {
