+++ spec/spec-helpers.js
@@ -1,0 +1,2 @@
+exports.ifit = (condition) => (condition ? it : it.skip)
+exports.ifdescribe = (condition) => (condition ? describe : describe.skip)
